Chore: Resolve compiler warnings
This commit addresses several compiler warnings to improve code quality and maintain C++98 compatibility.

- Resolved -Wunused-variable in tests/main.cpp and tests/test_ast_declarations.cpp by removing unused variables.
- Fixed -Wsign-compare in tests/test_lexer.cpp by adding a static_cast.
- Addressed C++11 extension warnings in src/bootstrap/source_manager.cpp by changing brace-initialization to constructor calls.
- Added explicit constructors to SourceLocation and Token structs to fix resulting compilation errors and ensure C++98 compliance.

+++ tests/test_lexer.cpp
@@ -207,7 +207,7 @@ TEST_FUNC(Lexer_ComprehensiveCrossGroup) {
         TOKEN_EOF
     };
 
-    for (int i = 0; i < sizeof(expected_tokens) / sizeof(TokenType); ++i) {
+    for (int i = 0; i < static_cast<int>(sizeof(expected_tokens) / sizeof(TokenType)); ++i) {
         Token t = lexer.nextToken();
         ASSERT_EQ(expected_tokens[i], t.type);
     }
